requirements.txt
Jinja2==2.10
pandas==0.23.4
python-dateutil==2.7.4
matplotlib==3.0.1
